Fix for blurry text in hovercard on iOS. UI elements will blur if set on non-integral X/Y position. Buddy = Ian H.

+++ ios/ios_src/WorldPinOnMap/View/WorldPinOnMapView.mm
@@ -171,8 +171,8 @@
     float trueX = x/m_pixelScale;
 
     CGRect f = self.frame;
-    f.origin.x = trueX - (f.size.width/2.f);
-    f.origin.y = trueY - (f.size.height);
+    f.origin.x = static_cast<int>(trueX - (f.size.width/2.f));
+    f.origin.y = static_cast<int>(trueY - (f.size.height));
     self.frame = f;
 }
